requirements.txt
chainlit
langchain
langchain-google-genai
python-dotenv
